convert PV power to float32

+++ power_perceiver/load_raw/data_sources/raw_pv_data_source.py
@@ -238,9 +238,10 @@ def _load_pv_power_watts_and_capacity_wp(
 
     # Load data in a way that will work in the cloud and locally:
     with fsspec.open(filename, mode="rb") as file:
-        pv_power_ds = xr.open_dataset(file, engine="h5netcdf").astype(np.float32)
+        pv_power_ds = xr.open_dataset(file, engine="h5netcdf")
         pv_capacity_wp = pv_power_ds.max().to_pandas().astype(np.float32)
         pv_power_watts = pv_power_ds.sel(datetime=slice(start_date, end_date)).to_dataframe()
+        pv_power_watts = pv_power_watts.astype(np.float32)
         del pv_power_ds
 
     pv_capacity_wp.index = [np.int32(col) for col in pv_capacity_wp.index]
